zsh: Rewrite git status prompt based on "porcelain" status
Rework the git status prompt to take advantage of the "porcelain"
git status option. The porcelain option causes git to output a
fixed-format machine-readable status. Using this feature simplifies
the code a bit and makes it easier to get various details about the
current status.

This should be a more robust implementation, no longer being
dependent on the specific wording of the status message. As well as
be easier to read and reason about.

This change also adds a couple more status hints to the prompt.