zsh: Fix tab completion causing malformed prompt on some systems
This problem seems to have been caused by incorrect locale variables
causing multi-byte unicode characters in my prompt being interpreted
as multiple characters when the tab completion system was redrawing
the prompt. Causing the string length calculation to be incorrect and
making completion be drawn a couple characters to the right of the
original position.

I found that this only happened on debian servers when logged in via
SSH. All of the LC_* variables were defaulted to "POSIX". Even though
the locale variable were set correctly when logged in directly.

This patch explicitly sets the LANG environmental variable to my
preferred locale (en_US.UTF-8) on shell startup.

An alternative zsh-based fix would be to wrap multi-byte unicode
characters in %{%G %}. But locale variables are important for other
reasons, so I'd rather have them set correctly.

There are several other ways to remedy this problem (such as the SSH
"SendEnv LANG LC_*" config). But I think that setting the variable as
part of the zsh setup scripts will be the most universal option in my
case. All of my systems have the en_US.UTF-8 locale available. And I
want all of them to use that locale unless I explicitly override it.

Helpful references:
 * https://stackoverflow.com/q/17931764
 * https://github.com/ohmyzsh/ohmyzsh/issues/7945
 * https://zsh.sourceforge.io/Doc/Release/Prompt-Expansion.html