zsh: Enable alternate screen terminal clear in systemd commands
The default value of $SYSTEMD_LESS is "FRSXMK". The less -X option
"disables sending the termcap initialization and deinitialization
strings to the terminal". Which prevents clearing the screen when the
program is closed.

This patch removes the -X option from the arguments that systemd will
pass to the less program when starting the pager.